Snapping Bigger Objects Added
Implemented ability snap all the faces/walls onto the playhouse. Made different types of walls snap to certain spots (e.g. the full size walls can only snap in one location on the sides of the playhouse[to prevent it from going into the ground] while the half-sized walls can snap to all the "wall" snap locations). Also made the slide snapable to the half sized wall that has a hole for the slide. Check out the scene: Snapping on Playhouse Test to check it out :)